% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{applicants}
\alias{applicants}
\title{applicants dataset}
\format{
A data frame with 49411 rows and 2 variables:
\describe{
\item{lens_id}{}
\item{applicants}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
applicants
}
\description{
A dataset of applicant names that have been cleaned in VantagePoint
based on a shared priority number and then shared inventor name.
The separator for names in the records is "|" not ";;" as in the raw
Lens data.
}
\keyword{datasets}
