% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{applicants_cleaned}
\alias{applicants_cleaned}
\title{applicants cleaned name dataset}
\format{
A data frame with 63243 rows and 2 variables:
\describe{
\item{lens_id}{}
\item{applicants_cleaned}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
applicants_cleaned
}
\description{
A dataset of cleaned applicant names. The names
have been separated onto individual rows per record.
}
\keyword{datasets}
