% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{applicants_ranked}
\alias{applicants_ranked}
\title{applicants ranked dataset}
\format{
A data frame with 31126 rows and 3 variables:
\describe{
\item{applicants}{}
\item{n}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
applicants_ranked
}
\description{
A dataset of cleaned applicant names. Names were cleaned in VantagePoint in
two steps using fuzzy matching on shared priority numbers followed by shared
inventor names.
}
\keyword{datasets}
