% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{citedpat}
\alias{citedpat}
\title{cited patents dataset}
\format{
A data frame with 27578 rows and 32 variables:
\describe{
\item{number}{}
\item{jurisdiction}{}
\item{kind}{}
\item{display_key}{}
\item{lens_id}{}
\item{publication_date}{}
\item{publication_year}{}
\item{application_number}{}
\item{application_date}{}
\item{priority_numbers}{}
\item{earliest_priority_date}{}
\item{title}{}
\item{abstract}{}
\item{applicants}{}
\item{inventors}{}
\item{owners}{}
\item{url}{}
\item{document_type}{}
\item{has_full_text}{}
\item{cites_patent_count}{}
\item{cited_by_patent_count}{}
\item{simple_family_size}{}
\item{extended_family_size}{}
\item{sequence_count}{}
\item{cpc_classifications}{}
\item{ipcr_classifications}{}
\item{us_classifications}{}
\item{npl_citation_count}{}
\item{npl_resolved_citation_count}{}
\item{npl_resolved_lens_ids}{}
\item{npl_resolved_external_ids}{}
\item{npl_citations}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
citedpat
}
\description{
A dataset of 27578 rows of patent documented cited in the core patent (pat) set.
}
\keyword{datasets}
