% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones_download.R
\name{drones_download}
\alias{drones_download}
\title{Download csv files for the Literature and Patent datasets}
\source{
The Lens Scholarly and Patent databases at
  \url{https://www.lens.org/}.
}
\usage{
drones_download(data = NULL, dest = NULL, conflicts = "overwrite")
}
\arguments{
\item{data}{One of "all", "lit" or "pat"}

\item{dest}{destination folder, cannot be null}

\item{conflicts}{Determines what happens if a file already exists. Choose
from "overwrite" (default), "skip" or "error".}
}
\value{
downloaded files
}
\description{
Download folders containing scientific publications and patent
  data for drones from the open science framework project folder
  \url{https://osf.io/jr87e/}.
}
\examples{
drones_download(data = "all", dest = "data", conflicts = "overwrite"),
drones_download(data = "lit", dest = "data", conflicts = "overwrite"),
drones_download(data = "pat", dest = "data", conflicts = "overwrite")
}
