% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{inventors}
\alias{inventors}
\title{raw inventor names}
\format{
A data frame with 258564 rows and 2 variables:
\describe{
\item{lens_id}{}
\item{inventors}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
inventors
}
\description{
A dataset of raw inventor names.
Use the lens_id for table joins.
}
\keyword{datasets}
