% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{inventors_cleaned}
\alias{inventors_cleaned}
\title{cleaned inventor names}
\format{
A data frame with 80065 rows and 3 variables:
\describe{
\item{lens_id}{}
\item{inventors_cleaned}{}
\item{records}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
inventors_cleaned
}
\description{
A dataset of cleaned inventor names.
Names were cleaned in VantagePoint using fuzzy matching and shared priority
numbers followed by a second step of matching on shared cleaned applicant
names.
}
\details{
Use the lens_id for table joins
}
\keyword{datasets}
