% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{ipc}
\alias{ipc}
\title{international patent classification (ipc) dataset}
\format{
A data frame with 187512 rows and 5 variables:
\describe{
\item{lens_id}{}
\item{ipcr_classifications}{}
\item{subclass}{}
\item{description}{}
\item{level}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
ipc
}
\description{
A dataset of international patent classification (IPC) data.
The IPC codes can be joined to other tables using the lens_id.
Short descriptions have been added at the subclass level to
assist with visualisation but may need further editing for length.
}
\keyword{datasets}
