% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{ipc_count}
\alias{ipc_count}
\title{count of international patent classification (ipc) dataset}
\format{
A data frame with 520 rows and 2 variables:
\describe{
\item{description}{}
\item{n}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
ipc_count
}
\description{
A dataset of international patent classification subclass codes with count.
}
\keyword{datasets}
