% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{lit}
\alias{lit}
\title{scientific literature dataset}
\format{
A data frame with 45266 rows and 29 variables:
\describe{
\item{lens_id}{}
\item{title}{}
\item{date_published}{}
\item{publication_year}{}
\item{publication_type}{}
\item{source_title}{}
\item{issns}{}
\item{publisher}{}
\item{source_country}{}
\item{authors}{}
\item{abstract}{}
\item{volume}{}
\item{issue_number}{}
\item{start_page}{}
\item{end_page}{}
\item{fields_of_study}{}
\item{keywords}{}
\item{mesh_terms}{}
\item{chemicals}{}
\item{funding}{}
\item{source_urls}{}
\item{external_url}{}
\item{pmid}{}
\item{doi}{}
\item{microsoft_academic_id}{}
\item{pmcid}{}
\item{citing_patents_count}{}
\item{references}{}
\item{citing_works_count}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
lit
}
\description{
Scientific publications mentioning drones from the Lens.
Access the public collection at \url{https://www.lens.org/lens/search/scholar/list?collectionId=199039}
}
\keyword{datasets}
