% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{lit_title}
\alias{lit_title}
\title{literature title dataset}
\format{
A data frame with 45266 rows and 2 variables:
\describe{
\item{text}{}
\item{doc_id}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
lit_title
}
\description{
Titles of scientific articles mentioning drones.
}
\keyword{datasets}
