% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{lit_title_nouns}
\alias{lit_title_nouns}
\title{noun phrases from the titles of scientific articles}
\format{
A data frame with 161760 rows and 6 variables:
\describe{
\item{text}{}
\item{doc_id}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
lit_title_nouns
}
\description{
Noun phrases extracted from the itles of scientific articles using spacyr.
}
\keyword{datasets}
