% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{pat_abstract}
\alias{pat_abstract}
\title{patent abstracts dataset}
\format{
A data frame with 49417 rows and 2 variables:
\describe{
\item{text}{}
\item{doc_id}{The lens_id}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
pat_abstract
}
\description{
A dataset of patent abstracts extracted from the patents file.
}
\keyword{datasets}
