% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{pat_title}
\alias{pat_title}
\title{patent title dataset}
\format{
A data frame with 49417 rows and 2 variables:
\describe{
\item{text}{}
\item{doc_id}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
pat_title
}
\description{
A dataset of patent titles extracted from the patents file.
}
\keyword{datasets}
