% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{pat_title_nouns}
\alias{pat_title_nouns}
\title{patent title noun phrases dataset}
\format{
A data frame with 140930 rows and 6 variables:
\describe{
\item{doc_id}{}
\item{text}{}
\item{root_text}{}
\item{start_id}{}
\item{root_id}{}
\item{length}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
pat_title_nouns
}
\description{
A dataset of noun phrases from patent titles extracted from the patents file.
Extracted using spacyr.
}
\keyword{datasets}
