% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{applicants}
\alias{applicants}
\title{applicants dataset}
\format{
A data frame with 31126 rows and 3 variables:
\describe{
\item{lens_id}{}
\item{applicants}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
applicants
}
\description{
A ranked dataset of cleaned applicant names.
}
\keyword{datasets}
