% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{applicants_cleaned}
\alias{applicants_cleaned}
\title{applicants cleaned name dataset}
\format{
A data frame with 31126 rows and 3 variables:
\describe{
\item{lens_id}{}
\item{applicants_cleaned}{}
\item{records}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
applicants_cleaned
}
\description{
A dataset of 31126 rows of cleaned applicant names with .
}
\keyword{datasets}
