% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drones.R
\docType{data}
\name{applicants_ranked}
\alias{applicants_ranked}
\title{applicants ranked dataset}
\format{
A data frame with 20176 rows and 2 variables:
\describe{
\item{applicants}{}
\item{n}{}
}
}
\source{
\url{https://www.lens.org/}
}
\usage{
applicants_ranked
}
\description{
The applicants cleaned dataset ranked by number of
records. The top result is blank (NA).
}
\keyword{datasets}
